Encode stage type fingerprint explicitly

+++ src/state_fingerprint.cpp
@@ -160,7 +160,7 @@ void AddStageFingerprint(FingerprintWriter& writer, const Stage& stage,
     if (stage.generation_seed.has_value()) {
         writer.AddPod(*stage.generation_seed);
     }
-    writer.AddPod(static_cast<int>(stage.stage_type));
+    writer.AddPod(static_cast<std::uint8_t>(stage.stage_type));
     writer.AddQuantizedFloat(stage.gravity);
     if (include_cache_generation) {
         writer.AddPod(stage.tile_change_generation);
